Fix TypeError: setValue(self, int): argument 1 has unexpected type 'float'

+++ SDEllipse_engine.py
@@ -41,7 +41,7 @@ class Worker(QtCore.QObject):
     relative to x / east.
     '''
     # Define the signals used to communicate
-    progress = QtCore.pyqtSignal(float)  # For reporting progress
+    progress = QtCore.pyqtSignal(int)  # For reporting progress
     status = QtCore.pyqtSignal(str)
     error = QtCore.pyqtSignal(str)
     # Signal for sending over the result:
@@ -290,7 +290,7 @@ class Worker(QtCore.QObject):
             percentage_new = (self.processed * 100) / self.feature_count
             if percentage_new > self.percentage:
                 self.percentage = percentage_new
-                self.progress.emit(self.percentage)
+                self.progress.emit(int(self.percentage))
 
     def kill(self):
         '''Kill the thread by setting the abort flag'''
